# Closest point performance improvement