--- a/plottable.d.ts
+++ b/plottable.d.ts
@@ -1251,7 +1251,7 @@ declare module Plottable {
         class Line extends Drawer {
             constructor(dataset: Dataset);
             protected _applyDefaultAttributes(selection: d3.Selection<any>): void;
-            selectionForIndex(index: number): d3.Selection<void>;
+            selectionForIndex(index: number): d3.Selection<any>;
         }
     }
 }
@@ -1262,7 +1262,7 @@ declare module Plottable {
         class Area extends Drawer {
             constructor(dataset: Dataset);
             protected _applyDefaultAttributes(selection: d3.Selection<any>): void;
-            selectionForIndex(index: number): d3.Selection<void>;
+            selectionForIndex(index: number): d3.Selection<any>;
         }
     }
 }