--- a/plottable.d.ts
+++ b/plottable.d.ts
@@ -330,11 +330,15 @@ declare module Plottable {
              */
             function setTimeout(f: Function, time: number, ...args: any[]): number;
             /**
-             * To be used in the first line of a deprecated method
+             * Sends a warning to the console. The warning includes the version number of the deprecation,
+             * the name of the function which was deprecated and an optional message.
+             *
+             * Tu be used in the first line of a deprecated method.
              *
              * @param {string} version The version when the tagged method became obsolete
+             * @param {string?} message Optional message to be shown with the warning
              */
-            function deprecated(version: string): void;
+            function deprecated(version: string, message?: string): void;
         }
     }
 }