--- a/plottable.d.ts
+++ b/plottable.d.ts
@@ -330,10 +330,10 @@ declare module Plottable {
              */
             function setTimeout(f: Function, time: number, ...args: any[]): number;
             /**
-             * Sends a warning to the console. The warning includes the version number of the deprecation,
+             * Sends a deprecation warning to the console. The warning includes the version number of the deprecation,
              * the name of the function which was deprecated and an optional message.
              *
-             * Tu be used in the first line of a deprecated method.
+             * To be used in the first line of a deprecated method.
              *
              * @param {string} version The version when the tagged method became obsolete
              * @param {string?} message Optional message to be shown with the warning