… src/drawers/drawer.ts
… src/drawers/lineDrawer.ts
… src/plots/barPlot.ts
… src/plots/linePlot.ts
… src/plots/plot.ts
… src/plots/scatterPlot.ts
… src/utils/windowUtils.ts
… test/testReference.ts
… test/tests.js
… test/utils/windowUtilsTests.ts