Closes #2195

Before: `660ms` http://jsfiddle.net/9emes3ud/2/ 
After: `1.5ms` http://jsfiddle.net/9emes3ud/3/
`600x` improvement

Stretching:
- 10 000 datapoints - `10ms` - http://jsfiddle.net/9emes3ud/4/ 
- 100 000 datapoints - `100ms` - http://jsfiddle.net/9emes3ud/5/ (mainly limited by the core rendering)

Bar Plot: 
- smooth: 100 bars - `1ms` - http://jsfiddle.net/9emes3ud/9/
- stretch: 5000 bars - `100ms` -  http://jsfiddle.net/9emes3ud/10/

---

Also as part of this PR: 
- Added `Utils.Window.deprecated()`
- deprecated `Plot._visibleOnPlot()`
- added `protected Plot._entityVisibleOnPlot(pixelPoint, datum, index, dataset)`

---

QE Notes: 
- Performance testing: closest points on Scatter Plot, Bar Plot, Line Plot should be improved 
- Regression pass: I changed the logic for retrieving closest points and for computing which data points are visible on the screen.