@bluong 2015-06-17
This pull request is certainly doing a lot at once.

It is:
* Improvements at the plot level through `_datumVisibleOnPlot()`
* Improvements at the drawer level through `_cachedSelection`
* `deprecated()` addition

How hard would it be to split this into a side branch and then having 3 separate prs?