@bluong 2015-06-17
Most of the PR looks fine aside from the idea of `LightweightPlotEntity`  I'm a bit unsure of the purpose of its existence