@aicioara 2015-06-18
`LightweightPlotEntity` is the core of the improvement. Basically, computing the `selection` for each entity was really time consuming. Therefore, we lazily compute it. That is why selection is set as null and only computed for the closest point. 

Now, I can see the source of confusion. The `LightweightPlotEntity` extents from `PlotEntity`, but that is not totally correct, as the `selection` field should be missing, and instead the `drawer` and `validDatumIndex` fields will be present (for later computation of `selector`). In the same time, I wanted this to inherit from `PlotEntity`, in case we want to add extra fields in the future.

Not really sure how to handle this.