@aicioara 2015-06-18
Writing some tests for the `deprecated()` method