@bluong 2015-06-18
The idea behind `_lightweightEntities()` seems really odd to me.  Judging from looking at the raw type definitions, `LightweightEntitiy` are actually heavier than the `Entity` type.  Also there is a lot of mysterious logic with the `selection` parameter being set to null first and then being set to a valid value later.