@aicioara 2015-06-19
It was weird that lightweight entities were inheriting from entities. 

The main reason behind doing so was such that the conversion back to a normal PlotEntity would be done by blacklisting extra fields instead of whitelisting the needed fields. The issue was that this bit of code would become fragile, but since the compiler is doing the type check, I believe it is a good compromise. Will do that