@aicioara 2015-06-22
I am in favor of making the `LightweightPlotEntity` interface public. That is because:
`entityNearest()` in BarPlot and LinePlot can also leverage a 10x extra increase by not calling `entities()` and instead calling the protected `lightweightEntities` (just like the other plots do). However, protected methods cannot use unexported types