@aicioara 2015-06-22
callee / caller are deprecated https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/arguments/callee